Renamed and changed the behaviour of copy file to move file so I don't have to delete the file afterwards.  Created createLinkCommand with tests

+++ DirLinker/trunk/JunctionPointer/Commands/DeleteFolderCommand.cs
@@ -17,8 +17,11 @@ namespace JunctionPointer.Commands
 
         public void Execute()
         {
-            _Folder.DeleteFolder();
-            _FolderDeleted = true;
+            if (_Folder.FolderExists())
+            {
+                _Folder.DeleteFolder();
+                _FolderDeleted = true;
+            }
         }
 
         public void Undo()
